+++ src/input_context/context_instance.rs
@@ -14,6 +14,7 @@ use bevy::{
 use super::{
     events::{ActionEvents, Canceled, Completed, Fired, Ongoing, Started},
     input_action::{Accumulation, ActionOutput, InputAction},
+    input_bind::{InputBind, InputBindModCond},
     input_condition::InputCondition,
     input_modifier::{negate::Negate, swizzle_axis::SwizzleAxis, InputModifier},
 };
@@ -572,64 +573,6 @@ pub enum ActionState {
     Ongoing,
     /// The condition has been met.
     Fired,
-}
-
-/// Associated input for [`ActionBind`].
-#[derive(Debug)]
-pub struct InputBind {
-    pub input: Input,
-    pub modifiers: Vec<Box<dyn InputModifier>>,
-    pub conditions: Vec<Box<dyn InputCondition>>,
-
-    /// Newly created mappings are ignored by default until until a zero
-    /// value is read for them.
-    ///
-    /// This prevents newly created contexts from reacting to currently
-    /// held inputs until they are released.
-    ignored: bool,
-}
-
-impl InputBind {
-    /// Creates a new instance without modifiers and conditions.
-    pub fn new(input: impl Into<Input>) -> Self {
-        Self {
-            input: input.into(),
-            modifiers: Default::default(),
-            conditions: Default::default(),
-            ignored: true,
-        }
-    }
-}
-
-impl<I: Into<Input>> From<I> for InputBind {
-    fn from(input: I) -> Self {
-        Self::new(input)
-    }
-}
-
-/// A trait to ergonomically add a modifier or condition to any type that can be converted into a binding.
-pub trait InputBindModCond {
-    /// Adds modifier.
-    #[must_use]
-    fn with_modifier(self, modifier: impl InputModifier) -> InputBind;
-
-    /// Adds condition.
-    #[must_use]
-    fn with_condition(self, condition: impl InputCondition) -> InputBind;
-}
-
-impl<T: Into<InputBind>> InputBindModCond for T {
-    fn with_modifier(self, modifier: impl InputModifier) -> InputBind {
-        let mut binding = self.into();
-        binding.modifiers.push(Box::new(modifier));
-        binding
-    }
-
-    fn with_condition(self, condition: impl InputCondition) -> InputBind {
-        let mut binding = self.into();
-        binding.conditions.push(Box::new(condition));
-        binding
-    }
 }
 
 /// Represents the side of a gamepad's analog stick.