+++ src/lib.rs
@@ -72,11 +72,11 @@ pub mod prelude {
         input::{GamepadDevice, Input, InputModKeys, ModKeys},
         input_context::{
             context_instance::{
-                ActionBind, ActionData, ActionState, ContextInstance, GamepadStick, InputBind,
-                InputBindModCond,
+                ActionBind, ActionData, ActionState, ContextInstance, GamepadStick,
             },
             events::*,
             input_action::{Accumulation, InputAction},
+            input_bind::{InputBind, InputBindModCond},
             input_condition::{
                 block_by::*, chord::*, condition_timer::*, hold::*, hold_and_release::*,
                 just_press::*, press::*, pulse::*, release::*, tap::*, ConditionKind,